[Redux] Move tags states from queriesSlice to tagsSlice (#5597)
* Update packages/desktop-client/src/queries/queriesSlice.ts


* [autofix.ci] apply automated fixes

* Move getAccountsById to accountsSlice

* [autofix.ci] apply automated fixes

* [Redux] Move tags states from queriesSlice to tagsSlice

* Fix lint + release notes

---------
