+++ packages/desktop-client/src/components/tags/TagEditor.tsx
@@ -3,14 +3,14 @@ import { type RefObject } from 'react';
 import { Button } from '@actual-app/components/button';
 import { ColorPicker } from '@actual-app/components/color-picker';
 
-import { type Tag } from 'loot-core/types/models';
+import { type TagEntity } from 'loot-core/types/models';
 
 import { useTagCSS } from '@desktop-client/hooks/useTagCSS';
-import { updateTag } from '@desktop-client/queries/queriesSlice';
 import { useDispatch } from '@desktop-client/redux';
+import { updateTag } from '@desktop-client/tags/tagsSlice';
 
 type TagEditorProps = {
-  tag: Tag;
+  tag: TagEntity;
   ref: RefObject<HTMLButtonElement | null>;
 };
