+++ packages/desktop-client/src/redux/mock.tsx
@@ -42,6 +42,10 @@ import {
   reducer as queriesSliceReducer,
 } from '@desktop-client/queries/queriesSlice';
 import {
+  name as tagsSliceName,
+  reducer as tagsSliceReducer,
+} from '@desktop-client/tags/tagsSlice';
+import {
   name as usersSliceName,
   reducer as usersSliceReducer,
 } from '@desktop-client/users/usersSlice';
@@ -56,6 +60,7 @@ const appReducer = combineReducers({
   [payeesSliceName]: payeesSliceReducer,
   [prefsSliceName]: prefsSliceReducer,
   [queriesSliceName]: queriesSliceReducer,
+  [tagsSliceName]: tagsSliceReducer,
   [usersSliceName]: usersSliceReducer,
 });
