+++ packages/loot-core/src/types/models/tags.ts
@@ -1,4 +1,4 @@
-export interface Tag {
+export interface TagEntity {
   id: string;
   tag: string;
   color?: string | null;
