+++ upcoming-release-notes/5597.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [joel-jeremy]
+---
+
+[Redux] Move tags states from queriesSlice to tagsSlice
